\name{createHdfsDirectory}
\alias{createHdfsDirectory}

\title{
Create a directory in hdfs
}
\description{
Creates the directory in the given parent directory with the dirName.
}
\usage{
createHdfsDirectory(parentDir, dirName)
}

\arguments{
  \item{parentDir}{
Location of the Parent directory where the directory is to be created
}
  \item{dirName}{
Name of the directory
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}



\seealso{

}
\examples{

createHdfsDirectory("/user/workspace/", "test")
}

