\name{deleteHdfsDirectory}
\alias{deleteHdfsDirectory}

\title{
delete the hdfs directory
}
\description{
delete the given hdfs directory from given parent directory
}
\usage{
deleteHdfsDirectory(parentDir, dirName)
}

\arguments{
  \item{parentDir}{
  Location of the Parent directory from where HDFS directory has to be removed

}
  \item{dirName}{
  Name of the directory

}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}


\seealso{

}
\examples{
#deleteHdfsDirectory("/home","/test")

}

