\name{get_password}
\alias{get_password}

\title{
Get the password from user
}
\description{
This function allows to get the masked password from user on console.
}
\usage{
get_password()
}

\details{

}
\value{

}
\references{

}
\author{
Prasanna_Teli@infosys.com
}
\note{

}

\seealso{

}
\examples{
#get_password()
}

