\name{IIP.encrypt}
\alias{IIP.encrypt}

\title{
Encrypt passwords
}
\description{
Function to encrypt passwords using key(RSA 2048bits) for encryption
}
\usage{
IIP.encrypt(pwd = NULL, pwdFile = NULL, encryptedFilePath = NULL)
}

\arguments{
  \item{pwd}{
Plaintext Original Password
}
  \item{pwdFile}{
Location of the file containing plaintext password. [Optional if the user is entering the password manually]
}
  \item{encryptedFilePath}{
Location of the file where the encrypted password should be stored.
}
}
\details{

}
\value{

}
\references{

}
\author{
Prasanna_Teli@infosys.com
}
\note{

}

\seealso{

}
\examples{
#IIP.encrypt()
}

