\name{IIPR-package}
\alias{IIPR-package}
\alias{IIPR}
\docType{package}
\title{
Package containing utilities for IIP
}
\description{
the IIPR package has been built to allow data scientists and analysts access views created in IIP
}
\details{
\tabular{ll}{
Package: \tab IIPR\cr
Type: \tab Package\cr
Version: \tab 1.2.4\cr
Date: \tab 2016-06-29\cr
License: \tab What license is it under?\cr
}

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{

}
