\name{readInputFile}
\alias{readInputFile}

\title{
Read the input file
}
\description{
This function read the file given at any hdfs location
}
\usage{
readInputFile(inputPath, readSep)
}

\arguments{
  \item{inputPath}{
the hdfs path to file
}
  \item{readSep}{
read seperator for file
}
}
\details{

}
\value{

}
\references{

}
\author{
KiranKumar_Kaipa@infosys.com
Poorva_Agrawal@infosys.com
}
\note{

}

\seealso{

}
\examples{
#readInputFile(inputPath="FileInputPath", readSep=",")

}

