\name{timeFormat}
\alias{timeFormat}

\title{
Convert given time into string
}
\description{
Get the timestamp as input and then convert it into a String, eliminating spaces and special characters.
}
\usage{
timeFormat(timestamp, myPattern)
}

\arguments{
  \item{timestamp}{
The time stamp to be appended to file name
}
\item{myPattern}{

}
}
\details{

}
\value{

}
\references{

}
\author{
Neethu_Prem@infosys.com
}
\note{
}


\seealso{

}
\examples{
#timeFormat(timestamp=TimeStampObject, myPattern="YYYY-MM-DD HH:MM:SS")
}

